Add WrappedText::count_lines method
This allows for counting lines without using the
`wrapped_text.lines().count()` method that would otherwise be necessary.

The information for the number of lines is already contained in the
breaklist of WrappedText, and it would be unnecessary and wasteful to
iterate over the lines---however cheap that may be---to simply get the
number of lines.