Implement basic flex system
To be honest, this commit is a bit more momentous than the name
suggests. This really is where the redesign is starting to pay off.

Only examples/minimal.rs shows of this functionality, for now, because
I'm eepy.

Note that examples/reader.rs is totally broken and
examples/status_display.rs does not have graphs anymore (temporarily)
since these have not been implemented under the new system, yet.