Fix bug in resizing WrappedText
Because the breaklist was not cleared before pushing the breaks as
calculated during rewrapping, iterating over its lines would read
outside of the breaklist, causing a panic.

The fix is rather simple, since it adds a call to clear on the breaklist
where it should have been there all along.

Also added a test that failed before and now passes.