Add examples/minimal_builder.rs
This is an illustration of the new builder pattern for the Element tree.

The Element tree that is set up in `examples/minimal.rs` and
`examples/minimal_builder.rs` are equivalent.

The previous way of creating the Element tree still works as before (as
illustrated by the unchanged `examples/minimal.rs`.
But, it is very likely that the old way will break once we start
relying on newer mechanisms such as that implemented in this commit.

References: https://todo.sr.ht/~ma3ke/stammer/7